The preceding bug concerning exp turned out to be more involved and affected all occurrences of choke.
For choke we have (and use) the following property:
  l <= choke (x, (l, u)) eps <= u

However, we do not necessarily have:
  l <= compress (choke (x, (l, u))) eps <= u
For example:
  compress (choke (10, (9, 11))) 100 = 0

Since bindR, mapCR and mapCR2 secretly perform compress on their arguments, this may result in incorrect results.

If we write:
  expUniformCts :: Integer -> Base :=> (Complete Base)

  realExpBound :: BoundedCReal -> CReal
  realExpBound a@(x,(Interval (_,ub))) = bindR (expUniformCts (ceiling ub)) (choke a)
there is secretly written something like:
  realExpBound a@(x,(Interval (_,ub))) = bind (expUniformCts (ceiling ub)) (compress (choke a))
Because expUniformCts does expect its second argument to be bound by the first, this can go wrong.

Therefore I let bindR, mapCR and mapCR2 perform compress on the output rather than the input.

Alternatively, we could let expUniformCts do a min (x, maxx). This is the way we do it in CoRN and how I do it in DReal.